Reduce - Initial Commit
I'm moving all the logic for reductions into a new library. What I have
right now is a helper function (mostly copy-paste from the existing
exhaustive test library) that can do general reductions recursively and
output everything it generates. I'll have to implement a user-level
function for reductions and then also subsets, and then choose a new
name for reductions, maybe 'contractions' or 'conversions'.

Because of the way I want to be able to customize and control the
testing, the implementation will now be done in the utility program,
pieced together from this library and maybe another library with some
tests. The additional function calls shouldn't cost too terribly much,
but we shall see.