Reduce - Range Relation, Recursive (bad) Subsets
Changed from the "outside the given M-range" criterion to either inside
or outside, since both are useful. Implemented a function for subsets,
which started out as single-order, only removing one value, but then I
realized it'd be good to have multiple value removal, and implemented
that the same as with the reduction function: recursively. But this is
so simple to do iteratively! So I'll go back and do it iteratively.