Reduce - Smarter Subsets
What I meant wasn't strictly 'iterative,' but rather I meant something
that would generate the subsets in-place, without doing excessive work.
The superset insertion algorithm uses some recursion for simplicity, and
so this does too. I still need to test it, and the method feels like it
could be a little shaky, but it's as clean as I could come up with,
doing as little work as possible.