Reduce - Testing Subsets, Fixes
Wrote up the test program for the Reduce library. It just does subsets
for now. But I noticed one little issue with subset generation, which
was that it'd still try to remove set elements when the set was already
at or below the target size. I also added input validation since the
function I was writing became the user-mode function. Finally, I also
fixed the test program for Expand.