Reduce - Contraction Testing, Fixes, Optimization
Put contractions (what I'm now calling the binary-operation reductions)
in the test program. There was an issue with hanging, which I identified
to be because of M-range restrictions preventing the loop from reaching
the increment. So I made it a for-loop and I think it's just more
intuitive that way anyways. Things seem to be working fine as far as I
can tell.

Did an optimization with subset production, where now it'll write out
the tail set values only once when there's one removal left, rather than
every time a subset hits zero removals. Also fixed input validation to
be non-descending rather than strictly ascending.