Tests - Library for Individual Set Testing
The General Exhaustive Test program has been reworked into a library for
individual (non-recursive) tests. I've added a test for successibility,
something that'll be used for testing unifiability if I ever want to.
Because of new developments in theory, it makes more sense to define the
intricate details of a test at the user-program level.

So I'll be reworking the reductive programs. This change has broken the
Weed utility, the current reductive testing program, so I'll need to
rewrite it. My plan is to give it more options for what reductive paths
to follow, and possibly add the option of loading a secondary record to
compare reductions to, memoizing results and/or getting a head-start on
creating data.