Top-Down - Renamed and Reworked
I've renamed the Weed program to Top-Down, to emphasize that it's taking
the reductive, or top-down method, which is one segment of our overall
approach to searching.

What used to be done automatically by the General Exhaustive Test,
besides enumerating contractions, is now done in here, allowing for more
configuration in the future. The top-down iterations still go down to a
base case of N = 4, and we still progressively check subsets in the same
way as we had before. The command-line arguments and options were left
unchanged for the time being.

The big new feature is the additional information stored in the records,
with new bits for bisectability. I'm anticipating we'll want to look
directly for set bisectability, both for the sake of data and for having
an even more ideal approach. The idea is that we can get all supersets
of bisectable sets by taking supersets of precarious sets (not
necessarily of one fewer element), and if a nullifiable set isn't a
superset, it's bisectable itself, so we'll mutate bisectable sets and
then do our top-down conversions to mark off the rest and complete the
dataset.

The reworked program seems to perform well, with no statistically
significant difference in runtime for the tests I ran up to N = 6,
though my workstation computer is running a CPU-intensive scientific
computing job at the moment, so I should probably test somewhere more
controlled. But it looks promising!

I had to go into the Reduce library to add a feature allowing the return
code of the output function to be saved and returned (by way of a simple
reference pointer). Otherwise we'd perform a test on the reduction and
lose the result immediately. I also had to fix it so that it actually
exits upon receiving a nonzero return code. This was only for the
Contraction function; I'll have to go back later and do Subsets.