Base-Up - Renamed and Reworked
Took the General Multiple Expand program and reworked it into what I'm
calling Base-Up now.

I made provisions for configuring what specific types of sets to use
expansion methods on. I've implemented them as a bit-mask to check
against specific bit-settings, but at the user level, it'll be something
like 'nullifiable', 'bisectable', or 'precarious' as the options.

The other thing is that I've implemented Mutations into the program. The
program will query the record once for all the superset expansions, and
then again for all the mutations. I actually found that, with
optimizations, this takes the same time as querying one time and having
the output function do sorting, which I'd say is more complicated. It
also takes the same time as the Generation program, as configured, for
the N = 6 test case I ran.

Next I want to work on how to implement these configurations at the user
level, for both this and Top-Down. We'll decide on a set of command-line
arguments. This will let us try out the various approaches we've cooked
up and test their relative speeds. Then we'll go down into the Expand
libraries and clean things up, merge things together into one.