Base-Up - User-Level Expansion Mode Controls
I've put in place some command-line arguments to give control over what
types of sets in the source undergo the expansion modes. Had to go and
implement a character parameter type in the argument reader function.

When I was testing things it seemed a bit weird, like sometimes when I
compiled it just would mark all destination sets. I was paying that
close attention to when I was saving and re-compiling between fixes
though, so it may be nothing.