Top-Down - 'Weak' Testing Idea
I've made a execution mode option for performing a 'Weak' test, which
essentially works as normal before, testing bisectability and subsets
and marking whichever is found first. By default, the program will test
strictly for bisectability, and it'll test all sets that haven't been
marked as fully tested for bisectability. I think this isn't that
helpful though, since we want precarious sets for most things, so we
should have it test only totally unmarked sets, not shown to be a
superset already.

I've also implemented an option to take first-order reductions to the
complement of the given M-range.

One thing that I theorized about was having to test additional
'reductions' that are first-order subsets (the set with one element
taken out). Something about "S_i is not in range" where S_i isn't
necessarily a contraction. I need to figure out what exactly I mean.