Evaluate, Top-Down - Focus on Precarious Sets
So I changed Top-Down to only test unmarked (uneliminated) sets by
default, and only test all bisectability-unknown sets when a command-
line option is given. Then I also went into Evaluate and added an
argument for the type of sets to display: innullifiable, nullifiable,
bisectable, or precarious. Now I can look at a listing of precarious
sets for the first time, and it's pretty interesting.

So the next thing I have to do is to combine the expansion libraries for
the sake of cleanup. Once everything is neat, then I think I'll work on
implementing additional record loading in Top-Down.