Expand, Base-Up - Merged Libraries Together
Merged the Multiple Expand library in with the Expand library, so it has
the Multiple Superset function. It was a direct copy-paste. Since now
the 'mutate' and 'supers' functions are at the user level, I had to add
input validation to the two functions, and change the names in the Base-
Up program. Seeing no inconsistencies.