Top-Down - Specifying Options
Since there's no real reason to perform reductive testing into a closed
M-range, I've re-specified it to be the M-range thoroughly expanded
from. The program will only test reductions landing outside any given M-
range.

Yesterday and today, I've underwent two different debugging sequences
because I was seeing bizarre results, like 338 or 366 innullifiable
length-6 sets 0-40, and the Top-Down program wasn't fixing anything. But
if I started over, it'd work. But problem wasn't the program, it was me.
I'd been trying to do Top-Down testing with other M-ranges, just as a
demonstration, and I'd forgotten about the 'tested for bisectability'
mark that the program would put down when it checked everything in that
specific M-range---the wrong, incomplete M-range. I'd even thought about
this before! Then the second time, I was trying to fix the original
record again with this knowledge in mind, but found it wouldn't mark off
any of the extra sets. But I'd known when I was making it that something
was off about my precarious set generation, that it was incomplete. The
numbers had been off. And I completely forgot to enable the subset
checking option every time I ran the program. So both cases were me
forgetting obvious side-effects of the good design of the program...
things I'd thought of. Renzl told me, "bro, you're actually seeing
ghosts." Yep. I was seeing ghosts.

So I implemented an option to re-test any sets that have already been
marked as non-bisectable, in case the wrong M-range is given or we want
to run a demo or something.