Base-Up - Progress Tracking
Did a copy-paste of the progress tracking stuff as implemented in the
Generation util. Did a cross-check with Top-Down and it matches up
everywhere it should. Maybe some of the common functions should be held
in the interface library? Not sure.

This was the command I used to test progress tracking on one of my runs,
with the fifo 'prog' for progress outputs:

    while true; do od -A n -t d8 -N 16 prog & killall -s USR1 baseUp;
            sleep 1; done

So here are the things I can think of doing. I should rewrite the
documentation blocks at the top of each program, and I should also add
more detail to the usage messages to clarify things. Then I need to go
and update the scripts. I'm thinking for the Auto script, I'll keep the
classic way of expanding things as an option, while also implementing
the 'precarious' strategy. Not much should need to be done with the
Database script, but I might put in progress tracking, with or without
the program-level feature. Finally, I can go in and fix the build system
and delete all the now-unused files, having successfully completed this
overhaul.