Database Script - Reworked for New Programs
Did some reworking for the new utility programs. Basically just fitting
what I have around the new names and new parameters. In the future, I
want to have the reductive process automatically figure out whether it
needs to be weak and what M-range to use, based on what's in the log
file. Besides that, it's pretty much all set.