Database Script - Inspection by Index Buckets
Added lexicographic index buckets to inspection. This happens
automatically, and the buckets are set up in the config file. Yeah, not
the best, but later we could maybe change it to a command-line option if
it feels better. It probably would be a lot better but we'll see.
Looking at the results there's definitely a big difference zooming out.