Evaluate - Bugfix, Bucket Optimization
The program would just fail when trying to do the index inspections.
It's because I didn't test my change with validating the list to be in
ascending order. Fixed now.

Also I thought about putting in a small optimization to save a bunch of
needless looping when sorting sets into buckets, by using a static
variable to keep track of the current bucket and advancing when we
surpass the cutoff. But it doesn't work with the way M-values are done,
so I'll just keep things the way they are.