Evaluate - Filtering Optimization
Put in the filtering optimization I was talking about. This requires
that M-value filters are in ascending order. But they're not cutoffs for
like buckets, they're still exact values.

I did this because it's just really annoying waiting two minutes for an
inspection to terminate with no indication of progress. It's all this
needless looping. My script isn't even smart enough to skip running the
command on records which don't even have the indices in their range. So
I'm thinking of making it smarter so it does that. I'm wanting to redo
inspections so I can control all the parameters in one go, like whether
I'm looking for innullifiable or precarious sets (or both, maybe?),
filtering by M-value, lexicographic index, or even just the record
number (or multiple?), and the range involved for the filters.

I also might bring back multithreading. I think I could solve everything
problematic before by just making separate arrays for each thread to put
filtering results in, then adding everything together at the very end.
So that might be it, but we'll see how well this optimization does.