Evaluate - Re-Implemented Multithreading
Multithreading is BACK. And it looks like this time it may actually make
a difference, since we can use it with filters. It seems to scale
decently, but since I'm running a process on the workstation right now I
can't quite tell. Hopefully this makes the waiting for inspection
results a lot less burdensome and annoying.