Database Script - Buckets all the way
Now the index-based inspection will generate buckets all the way down as
far as the end of the records. So providing a bucket size, it'll just go
as far down as possible. No more specifying the number of buckets. I had
to work it so that there aren't more than 1024 cutoffs at any one time,
so it has to iteratively invoke the command if there'd be more than 1024
buckets in one record file.

I may actually just leave it here, like it might not be necessary to
separate out different kinds of inspections. We'll see how well this
does for the analysis.