Database Script - Another Silly M-Range Mistake
I also forgot something when I implemented this M-range checker, and
that is fixed values being separate from the variable-segment M-range.
Gotta be careful! Anyways, I noticed in my mutative expansion process
that low record indices were being expanded into my destination DB
(which has a lower M-range) as expected, but then randomly index 0036
would be expanded. So yeah, that is because of this.