Database Script - Index Computation Bugfix
I've noticed an issue (ever since a while ago) where inspection into
index buckets will sometimes overcount the buckets needed for a record,
specifically when dealing with records with fixed values. Today I got
worried that it could indicate the index computation is wrong, but it's
not so bad. It's a stupid Python thing, because I'd been using the 'pop'
method on a list variable (function argument) and expecting it to not
modify the original. But of course it does. It's Python. This borked the
end-index calculation, and that's all. No real consequence besides more
numbers in the command. But now it won't do that.