Database Script - Logarithmic Bucket Base Options
Reworked logarithmic inspection buckets so that 2 isn't the only option
for the base. Now I have also the 4th-root of 2 so that there are four
buckets between every power of two. Nice, except for the round-off
errors causing the bucket cutoffs to be not quite multiples of 2. Oh
well. I could fix it with a little rounding.

Honestly I'm kind of worried about the bit that figures out which
buckets are needed for a given record. Round-off errors could get the
better of it. If I can't figure out a good solution, I might just extend
it by one in each direction since it doesn't slow down the eval program.